requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.0.2
openai>=1.0.0
mysql-connector-python
pymysql
sqlalchemy
python-dotenv
cryptography 